+++ src/tribol/mesh/CouplingScheme.hpp
@@ -128,7 +128,7 @@ class CouplingScheme {
      *
      * @return spatial dimension
      */
-    TRIBOL_HOST_DEVICE int spatialDimension() const { return static_cast<int>( m_mesh1.spatialDimension() ); }
+    TRIBOL_HOST_DEVICE int spatialDimension() const { return m_mesh1.spatialDimension(); }
 
     /**
      * @brief Return a view of the first mesh in the coupling scheme
@@ -1072,7 +1072,7 @@ TRIBOL_HOST_DEVICE inline bool CouplingScheme::Viewer::pruneMethodFacePair( cons
 
   auto& mesh1 = this->getMesh1View();
   auto& mesh2 = this->getMesh2View();
-  int dim = static_cast<int>( mesh1.spatialDimension() );
+  int dim = mesh1.spatialDimension();
   int num_nodes_face_1 = static_cast<int>( mesh1.numberOfNodesPerElement() );
   int num_nodes_face_2 = static_cast<int>( mesh2.numberOfNodesPerElement() );
